requirements.txt
Pillow
tqdm
humanize
colorama